choose function to utils
